lint fix for binary repr tests

+++ tests/arr_rs_tests/binary_test.rs
@@ -69,10 +69,10 @@ case(array_flat![1, 2, 3], array_flat![2, 2, 2, 2], Err(ArrayError::BroadcastSha
 
 #[rstest(
 num, expected,
-case(2 as u8, "10"),
-case(3 as u8, "11"),
-case(-3 as i8, "11111101"),
-case(255 as i32, "11111111"),
+case(2_u8, "10"),
+case(3_u8, "11"),
+case(-3_i8, "11111101"),
+case(255_i32, "11111111"),
 )] fn test_binary_repr<N: Numeric>(num: N, expected: &str) {
     assert_eq!(expected, Array::binary_repr(num))
 }
